Shards Shop(#19)
- return object instead of string for shop controllers
- Correct shop find by id in ShopService

+++ server/src/main/java/com/aion/server/database/entity/game/Shop.java
@@ -19,7 +19,7 @@ public class Shop {
     private String objectId;
 
     @Column(name = "item_id")
-    private int itemId;
+    private long itemId;
 
     @Column(name = "item_name")
     private String itemName;
